fix: refactor file structures.
