+++ projects/semantic_selector/adapter/one_to_one.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+# TODO
