+++ projects/tests/test_tokenizer.py
@@ -1,0 +1,7 @@
+import unittest
+from semantic_selector.tokenizer import InputTagTokenizer
+
+# TODO
+
+if __name__ == '__main__':
+    unittest.main()
